+++ kinto/plugins/default_bucket/__init__.py
@@ -96,9 +96,9 @@ def resource_create_object(request, resource_cls, uri):
     data = {"id": obj_id}
     try:
         obj = resource.model.create_object(data)
-    except UnicityError as e:
-        # The object already exists; skip running events
-        return e.object
+    except UnicityError:
+        # The record already exists; skip running events
+        return {}
 
     # Since the current request is not a resource (but a straight Service),
     # we simulate a request on a resource.