+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ This document describes changes between each past release.
 **Breaking changes**
 
 - Remove Python 3.5 support and upgrade to Python 3.6. (#1886)
+- Remove `record` from UnicityError class (#1919). This enabled us to fix #1545.
 - Storage backend API has changed, notions of collection and records were replaced
   by the generic terms *resource* and *object*.
 
@@ -18,11 +19,17 @@ This document describes changes between each past release.
 
 **Internal changes**
 
+- Remove depreciation warning for `mapping` (#1904)
 - Fix depreciated warn method (#1903)
 - Use f-string instead of % or format operators. (#1886)
 - Ignore admin plugin node_modules folder while running black (#1902)
 - Remove regexp py36 warnings. (#1907)
 - Changed psycopg2 dependency for psycopg2-binary. (#1905)
+
+
+11.2.1 (2018-12-09)
+-------------------
+
 - Still supports jsonschema 2.6 before 3.0 is released as a production release. (#1923)
 - Renamed core notions (ie. record and collection) (#710)